Implement `Formula::sat()`
This function is now implemented correctly, but all tests are failing
because `DerivationTerm` is not yet fully implemented for `Term`.  The
basic infrastructure for this example is now in place.

It seems like that, for `Derivation` we want a `DerivationQuery` rather
than an arbitrary function.  This will just make it easier to
instantiate, etc.  Also, we could have some kind of default which looks
for a single instance (i.e. `Term::Leaf(true)` in this case).