Big fix for `sat` example
It remains now to try and make this example much more efficient.
Ideally, we could read in some actual SAT problems in DIMACS form.  At
the current time of writing, there are a bunch of inefficiencies:

* Formula representation.  This is pretty inefficient as every clause
  requires its own allocated memory.  Also, variables are represented
  inefficiently.

* Domain size.  The `domain()` function is called on every step, and
  this algorithm is linear in the size of a formula.  This should be
  done offline.

Also, the derivation tree does not support other forms of optimisation.
For example, unit propagation.